add documentation for all backend side + swagger description

+++ backend/controller/register.js
@@ -5,6 +5,15 @@ const db = require("../util/DB");
 const loginController = require("../controller/login");
 const {validationResult} = require("express-validator");
 
+/**
+ * Function registers new user to the user table.
+ * It requires username and password fields in the request body: {username: "user login", password: "user password"}.
+ * All fields must be string-type.
+ * Function also prints helpful hints to server console in case of problems.
+ * @param req {object} request object from the previous function
+ * @param res {object} response object from the previous function
+ * @param next {function} next function
+ */
 exports.register = async (req, res, next) => {
     const errors = validationResult(req);
     if(errors.isEmpty()) {
@@ -37,6 +46,13 @@ exports.register = async (req, res, next) => {
     next();
 }
 
+/**
+ * The function delete a user account with all data related to it.
+ * The function requires jwt cookie (=user is logged in), and request body with the following data: {username: "user login", password: "user password"}.
+ * @param req {object} request object
+ * @param res {object} response object
+ * @param next {function} next function
+ */
 exports.deleteUser = async (req, res, next) => {
     const errors = validationResult(req);
     if(errors.isEmpty()) {
@@ -45,33 +61,22 @@ exports.deleteUser = async (req, res, next) => {
             const {username, password} = req.body;
 
             if (usernameCookie != null && username != null && password != null && username === usernameCookie) {
-                const selectUserQ = `SELECT *
-                                     FROM user
-                                     WHERE username = ?`;
+                const selectUserQ = `SELECT * FROM user WHERE username = ?`;
                 const selectUserResp = await db.makeQuery(selectUserQ, username);
                 const isUserExist = selectUserResp.length !== 0;
 
                 if (isUserExist) {
                     if (await bcrypt.compare(password, selectUserResp[0].password)) {
-                        const deleteJWTQ = `DELETE
-                                            FROM JWT
-                                            WHERE username = ?`;
+                        const deleteJWTQ = `DELETE FROM JWT WHERE username = ?`;
                         await db.makeQuery(deleteJWTQ, username);
 
-                        const deleteUserAllowedQ = `DELETE
-                                                    FROM UserAllowed
-                                                    WHERE username = ?`;
+                        const deleteUserAllowedQ = `DELETE FROM UserAllowed WHERE username = ?`;
                         await db.makeQuery(deleteUserAllowedQ, username);
 
-                        const deleteAccessRequestQ = `DELETE
-                                                      FROM AccessRequest
-                                                      WHERE sender = ?
-                                                         OR receiver = ?`;
+                        const deleteAccessRequestQ = `DELETE FROM AccessRequest WHERE sender = ? OR receiver = ?`;
                         await db.makeQuery(deleteAccessRequestQ, [username, username]);
 
-                        const userTablesQ = `SELECT *
-                                             FROM UserDatabase
-                                             WHERE username = ?`;
+                        const userTablesQ = `SELECT * FROM UserDatabase WHERE username = ?`;
                         const userTablesResp = await db.makeQuery(userTablesQ, username);
                         if (userTablesResp != null && userTablesResp.length > 0) {
                             for (let i = 0; i < userTablesResp.length; i++) {
@@ -81,22 +86,16 @@ exports.deleteUser = async (req, res, next) => {
                                 const tableDropQ = `DROP TABLE IF EXISTS ${tableName}`;
                                 const tableDropResp = await db.makeQuery(tableDropQ, tableName);
                                 if (tableDropResp) {
-                                    const deleteAllUserAllowedQ = `DELETE
-                                                                   FROM UserAllowed
-                                                                   WHERE id = ?`;
+                                    const deleteAllUserAllowedQ = `DELETE FROM UserAllowed WHERE id = ?`;
                                     await db.makeQuery(deleteAllUserAllowedQ, userTablesResp[i].id);
 
-                                    const deleteUserDatabaseQ = `DELETE
-                                                                 FROM UserDatabase
-                                                                 WHERE name = ?`;
+                                    const deleteUserDatabaseQ = `DELETE FROM UserDatabase WHERE name = ?`;
                                     await db.makeQuery(deleteUserDatabaseQ, name);
                                 }
                             }
                         }
 
-                        const deleteUserQ = `DELETE
-                                             FROM User
-                                             WHERE username = ?`;
+                        const deleteUserQ = `DELETE FROM User WHERE username = ?`;
                         await db.makeQuery(deleteUserQ, username);
 
                         await loginController.logout(req, res, next);
